move AccentColorOptions to a separate folder
