+++ frontend/src/metabase/lib/colors/types/accent-color-options.ts
@@ -1,0 +1,7 @@
+export interface AccentColorOptions {
+  main?: boolean;
+  light?: boolean;
+  dark?: boolean;
+  harmony?: boolean;
+  gray?: boolean;
+}
